Add function to compute area of simple polygon
And some base tests.
I'm experimenting a bit with how to best lay out these implementations. Let's try to test these things without a groper this time. However we also cannot use friend tests since that requires the user to bring in GTest in order to use Apex (even if not building the tests).