Add simple way to compare two simple polygons
This is basically a placeholder implementation exposing the implementation of std::vector.
To make this correct, we'd need to expand it to make it mark polygons with rotated vertices equal.
To make it perform in an OpenCL context, we'd need to make an OpenCL implementation for it as well. It can then be chained and compared on the GPU without needing to send it there.
All for later.