Implement CRTP to separate out protected implementation functions more neatly
This pattern essentially creates a template class that accepts a base typename that should match the footprint of SimplePolygon. It then has a function that casts its own instance to that SimplePolygon, in order to access its own members. This cast is possible because we let SimplePolygon inherit from the template class again. Weird construction, but effective.