Add new data structure for vector-of-vectors
This data structure will be optimised for sending to OpenCL as batch jobs. It contains all of the data in one contiguous array that can be sent directly to the GPU. However it will also need an index array that is sent separately, sadly. I didn't manage to fit that into the same array as it'll need to get stored in size_t alignments rather than the arbitrary Elements.