Remove the job and scheduler implementations
Too bad, but this approach won't work. It's impossible to determine beforehand how big the polygons will become after each operation. For instance, offsetting a polygon with N vertices and round joints with M vertices per circle can produce up to N*M/2 vertices in the final result. That sort of complexity is just unworkable if you want to chain multiple operations and reasonably schedule them. You can't predict how much memory is going to be used.
Instead our approach will be to execute each operation linearly, but still keep the batch polygons in memory on the compute device as long as the actual data is not necessary on the host. This should prevent copying data from and to the OpenCL devices as much as possible. It will also allow us to keep the implementation much simpler. Scheduling with memory constraints is an open problem in computer science, with some NP-hard problems defined for it.