Store end of used buffer in index as well
This is necessary because the polygons are not necessarily in order in the buffer, so we'd otherwise need to scan through all of the index buffer to see which polygon is at the end and how long it is.
Also this allows us to use the reserving constructor without assigning a segment of the buffer to each component polygon.