Add naive implementation for comparing batches with each other
This implementation simply compares if the data is saying the same. It'll work for now since we aren't going to be testing edge cases for this class.
In the future we'll need to use a true polygon comparison for this, which is a big algorithm that needs to be implemented in the GPU and such as well and needs to be moved to a different header with the CRTP. For now I just want to be able to use this in my tests to reduce code duplication there.