Use actual simple polygons to test batches
This is necessary since the test needs to test the by-reference effect of views on the batch. Because of this we can no longer just use the std::vector implementation that is our mock. So don't use the mock, use the actual stuff. It's more fragile to break if the simple polygon is implemented wrongly, but since that implementation should be transparent the code is very simple and not likely to break. Instead we now get proper errors if the batch is implemented wrongly, which is more doubtful.