Require at least GCC8
It seems that GCC7 still has a bug in it that causes the non-const overloads of functions to get selected sometimes even if the object is const. I could find some bugs in the bug tracker that seemed like they could be what I'm running into but I'm not sure if those were really it.