Adjust documentation to match new plan of storing indices
Since iterators become pretty much incompatible with the idea of constructing a view upon accessing the batch, we're going to need to store the views permanently. Instead of the index buffer, we'll store a list of views. This has one major drawback: Each view will need to use 4 pointer sizes of memory (start pos, end pos, capacity and a pointer to the batch). So we're going to have 33% more overhead. Another minor drawback is that we can no longer send the size and the position to start the next polygon to OpenCL, since those are not included in a pure list of views (but they were in the index buffer). The size can be inferred from the size of the view buffer so that's not a problem, but finding the position to store the next polygon is linear. However I expect that due to the parallel nature in OpenCL this information would be pretty useless anyway.