Add method to access internal storage data structure
This is necessary in order to copy simple polygons properly when placed in a batch. The alternative is a friend class, but I don't want the batch to really depend on the implementation of the simple polygon so I'll make this part of the API of the simple polygon instead.