Detect difference between version 1.7 and 1.8
In this version they stopped documenting changes. Very annoying.
I found that they added injection points for several definitions. These are global with macros so they can be detected. This test will be a bit fragile if they decide to rename them later, because I'm sure that these are not supposed to be part of the public API of this library and so they are not guaranteed to stay in place w.r.t. API breaks.