Require version 1.10 of GoogleTest
This is what we did it all for. We now have proper version detection on GoogleTest in CMake. So if the user uses a too old version of GoogleTest we'll now be able to give him an early error in CMake rather than having him find that out when he gets a compilation error and then having to figure out that the compilation error is due to having too old a version of GoogleTest which is not documented anywhere.