Don't name area function header specific to simple_polygon
We're going to need a bunch of very similar implementations of computing the area. One will be to compute the area of a simple_polygon (aggregate all vertices of a vector to a single area). One will be to compute the area of a complex polygon (aggregate all vertices of multiple vectors to a single area). One will be to compute the area of a batch of simple polygons (aggregate all vertices of a set of vectors to a matching set of areas). And one will be to compute the area of a batch of complex polygons (aggregate all vertices of multiple vectors of vectors to a matching set of areas). Each of these will be slightly different, but likely share the overall behaviour and theory. For this reason, I'd like to put all of them together even though they will be functions of various different classes.