Document usage of OpenMP for parallel computation rather than Boost.Compute
I had to make a hard choice here. I've tried with OpenCL, but support for that is really in limbo for almost all platforms. Intel pushes for it, but Intel produces rather weak parallelism (with their fake GPU-in-chip thing) and I don't use Intel. AMD hardly supports it at all. And NVidia supports only version 1.1 and even that is spotty. I've tried using Boost.Compute as a way to wrap around all the bugs of OpenCL, letting the Boost developers fix those OpenCL support issues for me. But Compute was last modified in 2018 so the project appears to be rather dead and support for newer devices is unknown.
OpenMP is a completely different paradigm, and they don't support heterogeneous computing. But they do support massively parallel computing on the GPU. Since I'm building this strategy pattern implementation with Nadir anyway, that is really all I need. I can choose myself whether to offload to GPU, to do normal parallelisation in the CPU, or to simply use a single-threaded implementation. No need to have the heterogeneous implementation then. The target is a single computer anyway.