Implement version of calculating polygon area that runs on the GPU
It seems that the overhead is just too big for this one. It's never faster than the multi-threaded approach on my computer, anyway. It kind of makes sense, because sending the data over to the GPU still needs to be done by the main processor, and probably needs more calculations than just a few multiplications and sums.