Test all implementations of area calculation
I've already locally tested them with these unit tests, but now it's truly keeping all of them tested by testing all variations automatically.
I've considered automating this further, like maybe generating a parameterised test where the function-under-test is a parameter. But that becomes really complex really quickly, which doesn't help to keep the tests maintainable. This is a simple approach which is fine for const functions like this.