Don't specify a particular compiler in CMake
Previously this was necessary because the default compiler in the build server provided by Github was too old.
However the specification of ubuntu-latest has changed. Ubuntu-latest is now a newer version of Ubuntu. This newer version no longer has g++-8, so the test fails. Let's try without specifying the compiler...