Add probe to see if GPU acceleration is properly supported
Only build the GPU-tests if GPU is properly supported. The tests don't look at which solution is the most efficient for the current problem (which would not consider GPU code if there is no GPU) but will test all solutions. Unfortunately that would mean that the GPU tests would be ran even if the testing computer doesn't have proper support for running things on a GPU accelerator.
This way we can test whether the build server has support for GPU acceleration, and only then build the tests that involve the GPU acceleration. That way instead of failing the tests, we'll skip them if no GPU is available.