Map vertices rather than whole polygon
The polygon is a non-trivially-copyable type, and as such isn't guaranteed to work well when mapped to the GPU. In fact, the GPU will then most likely not work on the correct vertex data.
Working on the vertex data itself guarantees that the vertex data gets copied to the GPU correctly.