Combine parallel computation with SIMD
These calculations can be done in parallel since there are no interdependencies. Although Clang seems to automatically detect this anyway. I measured the performance gain from this at 2%, which is within the 3-sigma noise in my measurements, so I'm not sure if it has any real benefits. Can't hurt anyway.