Remove GPU sync state tracking from SimplePolygon
We might need to do something clever with the updating of this state if an external user makes a lot of modifications vs. if we make modifications from the library's own algorithms. If the user of the library makes modifications we want it to behave consistently even if some data was secretly still on the GPU, so we'll need to sync. If we make our own changes, we just want to be as fast as possible and we can guarantee synchronisation ourselves. We'll have to have a good think about those things...