Implement most of SimplePolygon by turning it into a batch of points
The batch of points is safe to inherit from, so this is fine.
The SimplePolygon needs to behave precisely like a vector of points. Now it so happens that we already have a class that does precisely that: Batch.
Because we're inheriting from the template argument, calling functions of the base class from the subclass requires full specification of which class to call the function from: The template argument's function. The dispatch will be derived then during compile-time in the second stage. Harder on the compiler, and it requires some extra things to know when writing that code. However that is only for the library maintainers, not for library users. And the compiler gives an error when you make a mistake.

Pretty happy with how much this reduces code duplication.