Remove SimplePolygonBatch implementation
Use Batch<SimplePolygon> instead. That's what it's made for.
This also removes the tests, which had a bunch of tests for the batch of simple polygons which I haven't ported yet. However I think the new tests that we'll need for the batch-of-batches implementation will not involve anything that is dependent on SimplePolygon, to keep the test more separated.