Move Batch-of-Batches implementation to BatchBase
This is possible, because we now know that we can inherit from vector safely. So since we no longer need to inherit from BatchBase<SubbatchView>, we can actually *be* BatchBase itself. Previously, splitting up the Batch from the BatchBase class was kind of a workaround for not being able to inherit from other class-specialisations of the same class. The workaround is no longer needed because we inherit from vector instead.

This is also necessary for the future, because we need to implement a batch of simple polygons, which will inherit from BatchBase again and extend its behaviour. Since Batch<SimplePolygon> can't inherit from Batch<Batch<Point2>> (since they are the same class) we need to have the implementation for batch-of-batches in a separate class. We can do that with BatchBase now.

The new class privately (or rather, protectedly) inherits from vector, but it doesn't pass on all of vector's functions yet. Some of these needed to be implemented for the Batch-of-Batches anyway since they transform the parameter types, e.g. the copy constructor making copies of BatchBase instead of copies of vector. This was still under development before this move, and will need to be finished after.