Pass on getter functions from vector to BatchBase
These are the functions we can directly take over from std::vector, as they don't require conversion from Batch instances as elements of the Batch to the actual elements of the Batch, the SubbatchViews. The Batch of Batches returns SubbatchViews instead of the elements you put in, so that is a bit weird for the user but since they behave the same it's acceptable. But the rest of the functions will need some conversion somewhere.