Implement emplace for batches of batches
This one was quite a let-down. Since the batch of batches must behave externally like you can place normal batches in this batch, the user would expect to use the constructor arguments of BatchBase here. However the std::vector won't provide that since the constructor of its elements is actually the constructor of SubbatchView, which is completely different in its signature.
We also don't want to actually construct a new BatchBase. This would be wasteful since it would allocate new memory and write everything there. After that we would have to copy all of that data, resulting in more than twice the runtime.

Instead we'll have to re-implement all of BatchBase's constructors so that all of them exist. It's different overloads then, which is fine for the user, and we can even document it a bit better.