Add test for clearing the batch of batches
The behaviour in this function is really just calling the parent class (vector) to clear it, but does some additional logic to make sure subsequent modifications to the batch still work correctly. However since this is a unit test that just tests the clearing of the batch, this is not tested. Perhaps that would be a reason to make some basic integration tests too, alongside the fuzz test that I also have planned to make sure the logical behaviour is identical to vectors of vectors.