Guarantee a minimum capacity of 1 for every subbatch
This fixes an issue with our new optimisation where we don't move the array if it is the last subbatch. The detection of whether a subbatch was the last depended on the starting position of the subbatch and its capacity summing up to the next position. If the capacity of multiple subbatches at the end was 0, then all of them would be considered as the last one. So to prevent that, we have to ensure that all subbatches have a capacity of at least 1.
This is a performance hit if the batch contains a lot of empty subbatches, but a pretty minor one. Considering that the batch metadata is already 256 bits, one extra subelement is likely not going to make much of a difference. I consider the use case of empty subbatches to be very small. On the other hand, the performance gain of not having to reallocate the last subbatch is pretty substantial for the common use case.