Add constructor to convert subbatches to normal batches
This prevents ambiguous overloads, e.g. when taking an initialiser list with empty brackets which could be a real batch or a subbatch, while still allowing to use mixed subbatches and batches. It does create a copy, which may be elided by the compiler. Time will tell if this can be maintained.