Fix const limitation caused by const_iterator
We don't need to change directly from the iterator. We can look up its position from the start and just index from there.
However this also indicates an underlying problem that the iterator may be invalidated by the reallocation done by the insert operation on the vector of SubbatchViews. Basically after any insert or emplace operation, the iterator may be invalidated. Weird, but simple enough to work around.