Simplify benchmarking of area function
Less structure. Perhaps more repetition in the future. But we're going to need to customise lots of components here, including the part of constructing a SimplePolygon of the correct size, which shouldn't be included in the time measurement.
So let's just do this the simple way for now. It works.