Add free functions to calculate area of simple polygons
We're going to move away from the CRTP method. It is a bit clunky compared to the free functions, and is leading to collisions in the inheritance, a big template hell that I don't know my way out of. And if I don't know my way out now, future me will certainly be even more confused. As well as future other people.