Add free functions to translate simple polygons
Similar to what we just did with apex::area, we want to get rid of this CRTP which is a headache of template classes and inheritance. Instead we'll use free functions, and add convenience functions from the SimplePolygon class itself to make it easier for the user to discover how to call these free functions.