Replace string loader with include loading mechanism
This is different. Not traditional at all. But it is handy. This way, the strings get packed into the binary, which prevents needing to specify the correct path to the test data files.
Perhaps it'll be untenable once there is a lot of test data. We'll see about that then. Cross that bridge when we get there.