Define SVGs directly as std::string instead of char*
This prevents the need for converting.
It is no longer constexpr now. The compiler has to figure out whether this prevents the need to copy. I don't know if this is guaranteed, but it's likely.
When Clang supports constexpr strings we can turn this back into constexpr for better guarantees.