Fix finding coordinate end point
We need to start searching from the current position, not from the start of the file.
Also mark the closing quotes as end position, otherwise you might get the rest of the file also parsed. This would automatically work through the stroll function, but it's neater to actually just have the coordinate in the coordinate_str.