More robust parsing with malformed files
At least check if you actually got a position there, bud.
Still not robust against floating point numbers though. That is a bit more difficult because stod would cause rounding errors at the high numbers that we want to be able to put in these SVGs. We want to be able to use the coord_t maximum and minimum.