Directly store simple polygons as simple polygons
That way we don't have to re-generate them for each test. They are const anyway. And what's more, we then also don't have to know how to construct the polygons properly. And it works the same as the procedurally generated ones, too. Nicer in every aspect if you ask me.