Merge generators into test_simple_polygons
Now all of the simple polygon test cases are in one file.
This means that we won't have all of the test_case_loaders in one file any more. But that also kind of makes sense. You don't want the loaders for a polyhedron to be in the same file as for a polygon, since they have vastly different functions and use cases.