Add method to load a batch of simple polygons from SVG files
This works similarly to the single simple polygon, except it keeps parsing for more <polygon> tags and adds them all to a batch. There is a bit of duplicate code here, but since it's not all that much and a bit interweaved, it would make the code less readable if I were to factor it out.