Remove mocks for simple polygons and their batches
These mocks were not tenable. The main problem was that I couldn't realistically mock the underlying structure of the batch of batches. I needed to use the subelement array for testing functions on the GPU, because those GPU functions use that subelement array in their functions. That's the actual whole point of having that array in the first place. However to mock having all subelements of a vector in a single array and have it properly update the actual vectors too was a bit complex for a mocking class. I'd essentially have to build the whole thing again or use the real class (which has its own extensive test suite). I chose to use the real class, for better maintainability as long as it doesn't fail on us.