Put simple polygon test cases in a separate class
Our tests are currently not performing well. They take a lot of time to execute. The cause of that seems to be these globals. They seem to get re-initialised with every test. Not just every executable, but actually every single test. Because of the 1-million vertex circle this takes ~0.4s on my computer, for each of the dozens of area and translate tests. And these tests are about to become much more numerous too.
To prevent that, we're going to have to create a fixture. And not just a fixture specific to a bunch of tests, but actually a global fixture that we can use everywhere. Each testing application will have to re-initialise them, but that's okay as long as it's not every single test.
GoogleTest has a TestEnvironment option, but that requires defining a custom main which breaks the test detection feature in CMake. So I'm kind of doing that manually with this static global in each compilation unit.