Move batch test cases into a class too
Same as for the simple polygons themselves.
However it seems that this didn't completely solve the issue. The static field is still re-initialised.
I'm starting to suspect that this is not preventable because CMake calls upon the test application separately for each individual test, each of the 180. Best we could do then is to minimise what we need for each of the tests. We could try lazy-loading or to have each of these test cases be a functor to return a new instance rather than try to store them all in memory.