Rename SimplePolygon to Polygon
Never thought it would come to this.
The class was just named badly. All geometric definitions of a simple polygon say that it's a polygon that doesn't intersect itself. That is just not what this class represents. This is a polygon in the most basic definition of it: Just a set of edges/vertices that form a loop. It can self-intersect and everyone needs to know that it can. I have properties now that say whether or not it self-intersects.
The original name of the simple polygon came about because of the multi-polygons that can have multiple loops (e.g. consisting of multiple disparate parts or having holes in them). I'll just give those a different name, like MultiPolygon.