Fix warnings in documentation generation
I didn't intend to actually link to the ::apex::detail namespace. This is a namespace and doesn't really have any documentation. It's not in any specific file either. Linking to it doesn't really make a lot of sense.
We also can't link to std::vector. It's not in this project. We could link to a website, but let's not.