Reset polygon properties wen emplacing a vertex
We really don't know anything any more then. If convex or degenerate it could become concave. If concave it could become degenerate. If self-intersecting it could go around the intersection. If not intersecting it could become intersecting. If rotating either way it could become a mixed rotation, or if it is mixed it could become rotating in either direction.