Return optional intersection point instead of whether it intersects
We'll return more information if an intersection has been found. This is easy because once we've found whether they intersect, the intersection point is really already known. The caller will most likely need it so we might as well just return it already.

This does require making very clear what point gets returned if the intersection is a line segment itself (insead of just a single point).