Add documentation for calculating orientation of a point w.r.t. a line
We'll still need to make all of these implementations, but I'm honing in on the solution here. These are all going to be subroutines of each other. Really what we're doing here is laying down a mathematical framework of useful helper functions that will help us out a lot later on.