Make a simpler implementation of orientation with line
Let's not start over-engineering too much here. Yes, it would be a neater API if there were a function that returns an enum of whether the point is left, right or on the line. However this is very hot code, and constructing that enum would take quite a lot more time. In the end, it's not all that much different for the caller to interpret the result here. Instead of checking the enum values they'd have to use greater-than or less-than. Simple.