Implement line segment intersection
Took me like a week to figure out how to do this efficiently, but this is about as good as it gets. There's 18 integer operations, 4 comparisons and 2 divisions. That's lightning speed.
I'm not entirely happy with the interface of this function. There should perhaps be a boolean output version that just checks if they intersect, which is the only thing that's relevant for e.g. checking if a polygon has self-intersections.