Fix range-checking if divisor is negative
If the divisor is negative, then the parameters still need to be between 0 and the divisor. But then the check '< 0' would not indicate that it's outside of the range. The entire range is in the negatives, after all. In essence, '0 < p < divisor' can't be true because '0 < divisor' is not true.
This solution requires a bit more calculation, sadly. But hopefully still cheaper than actually doing the division in float math.