Vertex-intersections should be reported for both endpoints
This changes the expected behaviour. It is already implemented this way, but we'll now make this the expected behaviour so the documentation is changed to reflect that.
I want to keep it this way, for several reasons:
- For an external user, it is more logical that the intersection result is the same regardless of how the line segments are oriented.
- It is faster to check this with partially overlapping (colinear) line segments, resulting in fewer cases to check and fewer branches.
- It gives more information for self-intersections; both segments adjacent to a vertex intersection are reported to be intersecting.
- It properly gives an intersection for two line segments if they intersect in their ending endpoints.
This is not a problem for the self-intersection detection algorithm. It would normally give an intersection for every vertex (because its two adjacent segments would intersect) but this needs to be special-cased anyway because they could only intersect by overlapping, and so special-casing can improve performance significantly.