Make cross product also constexpr
For one, because we can. We should do that more often.
But also because this breaks on GCC when the constexpr orientation_with_line calls it. I'm guessing that Clang either inlines it such that it becomes constexpr anyway, or that Clang only evaluates it in the constexpr case and then also makes it constexpr or something.