More efficient self-intersection check of neighbouring elements
Two adjacent edges of a polygon can never truly intersect. We are guaranteed that they share a vertex. At worst, it can walk back over itself, creating a self-intersection by having two of its edges partially overlap.
We don't need to do the whole line segment intersection here. Something much simpler can be done: Simply check if they are collinear (one cross-product and subtractions) and then check if they are collinear in the same direction. If they are, they overlap and we have a self-intersection. This should add a slight bit of performance.