Replace cross products with cross product function
A few of these were actually cross products (as a result of the system of equations) but not marked as such. This makes it a bit more clear and concise. It's also slightly more magic, because the documentation doesn't mention it as a cross product, but if one knows what a 2D cross product is then it is hopefully quite clear, and might help bridge this formula to other intersection techniques.