Allow comparing self-intersections
I could've used the default here as well, but that involves checking if the position is also the same. That should be unnecessary, since these intersections are only valid as long as the polygon is unchanged anyway.
This is necessary to validate the results in the tests. But I'm sure it could be useful to the user as well.