Fix crash due to integer underflow at index 0
It was doing size_t - 1, which causes the integer to underflow to 18 quintillion. The array then quickly reaches out of bounds, resulting in a segfault.

So we should change that comparison to only add. That way we'll never get an underflow. Only an overflow, but that only happens when you have a vector with actually 18 quintillion vertices which is... unpractical anyway.