Fix intersection of 2-vertex polygons
This intersection should only yield one intersection point. However we were checking for each segment whether it intersects with the previous one. Both segments in the polygon would say that they intersect with the previous one, because due to having only 2 segments they were checking the same pair of segments. So let's special-case that.