Fix mirroring B dimension
I made a logic mistake here when I wrote this initially, a few months ago, I think. The resulting B coordinate needs to range from ceil(N/2) to N-1 for a polygon of size N. And as such, the B coordinate simply needs to be size() - 1 - x here. I think originally I was thinking in coordinates of the table above, which don't exactly match the vertex indices.