Add beginnings of a pairing function definition
We have this in our self_intersections function and we'll want to re-use that from the multi-threaded implementation in the GPU implementation too. So let's factor that out for re-use.
It'll need two functions. One to get the number of pairs, and another to get the individual pairs. But this is a start. Also needs some good tests, but that'll be quite easy.