Fix number of pairings with 0 elements
I tried quite hard to prevent there being a split in the code paths here, but couldn't come up with a solution. Too bad. So far it seems like we won't need to compute this in a hot loop luckily, but who knows what this will be used for in the future.