Don't crash getting pairs in sets of 1
Otherwise it would divide by N-1, and in a set of 1, N-1=0. Division by zero causes it to crash.
This immediately also catches the case of N=0, which would previously cause an integer underflow causing huge numbers to be output. Not important since we're not defining what the output should be in such cases, but since we're doing that check anyway we might as well catch it.