Fix using GPU definitions in benchmarker
Either we add the ifdef things to the benchmarking code and add the GPU detection code to CMake (which would be nicer in the long run), or we just always add the GPU definition and don't modify the benchmarking code.
For now I'm just adding this here. If we ever want to benchmark the CPU on a computer that doesn't have the GPU support, then we can still modify the benchmarking code for that.