Make a start towards a GPU implementation of naive self-intersections
Only the first half of it for now, the part that creates the position index.
The position index is a little bit different this time. It tracks the number of duplicates in a row, rather than the number of different coordinates encountered so far. Tracking the number of duplicates in a row is a bit friendlier to GPUs since we can't split up the array into an equal number of parts depending on how many threads we have.