Fix case where all vertices are the same
We'd never identify the loop start then, so we'd never properly count the number of identical vertices.
When we find out that all vertices are the same, we know that there are no proper intersections so we can immediately return. We return with RVO though so it's not an actual return.